requirements.txt
# Core dependencies
vosk>=0.3.45
sounddevice>=0.4.6
numpy>=1.21.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Optional enhancements
watchdog>=2.1.9      # File system monitoring
difflib              # Built-in fuzzy matching

# Development tools
pytest>=7.0.0        # Testing framework